poster_generator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pillow==10.1.0
python-dotenv==1.0.0
langchain-ollama==0.1.0
langchain-core==0.1.0
requests==2.31.0
websocket-client==1.6.4